moved publishing strategy to client package
