--- asciidoc-confluence-publisher-client/src/main/java/org/sahli/asciidoc/confluence/publisher/client/metadata/PublishingStrategy.java
@@ -1,11 +1,0 @@
-package org.sahli.asciidoc.confluence.publisher.client.metadata;
-
-/**
- * @author Laurent Verbruggen
- */
-public enum PublishingStrategy {
-
-    APPEND_TO_ANCESTOR,
-    REPLACE_ANCESTOR
-
-}
